+++ backend/app.py
@@ -1,13 +1,31 @@
 import os
 from flask import Flask, jsonify
+from pyspark import SparkContext, SparkConf
+from pyspark.sql import SQLContext
 from query_generator import query_builder
 
 
 app = Flask(__name__)
 
+conf = SparkConf().setAppName('Spatial-Crime-Analysis')
+sc = SparkContext(conf=conf)
+sqlContext = SQLContext(sc)
+csv_file = '../datasets/example.csv'
+df = sqlContext.read.format('com.databricks.spark.csv').options(header='true', inferschema='true').load(csv_file).\
+    selectExpr(
+    "RatecodeID",
+    "VendorID",
+    "passenger_count",
+    "DATE(CAST(tpep_pickup_datetime AS timestamp)) AS date",
+    "pickup_longitude",
+    "pickup_latitude",
+    "dropoff_longitude",
+    "dropoff_latitude"
+    ).where("date between '2016-01-01' AND '2016-01-05'")
+
 
 @app.route('/run_sample', methods=['GET'])
-def table():
+def query():
     try:
         result = query_builder.sample_query(df)
         return result
@@ -17,7 +35,6 @@ def table():
 
 if __name__ == '__main__':
     try:
-        query_builder.configure_spark()
         app.run(port=os.environ.get('FLASK_PORT', 8080), host='0.0.0.0')
     finally:
         sc.stop()
